Revert change of high_boart_::ply from uint32 to uint16
Doing this cost a few bytes in mid.wasm, and I don't want to do that.
So mid.wasm is back at 12,407 bytes.  Unfortunately this does slightly
increase the size of mid_worker.js, from 889 to 933 bytes.

+++ pentago/high/board.h
@@ -19,7 +19,7 @@ using std::tuple;
 
 class high_board_t {
   uint64_t side_[2];  // black, white (first player, second player)
-  uint16_t ply_;
+  uint32_t ply_;
 
   high_board_t(const side_t side0, const side_t side1, const int ply) : side_{side0, side1}, ply_(ply) {}
 public:
